# docs/swagger.yaml
basePath: /
definitions:
...
  slots.ErrorResponse:
    properties:
      message:
        description: Сообщение об ошибке
        type: string
    type: object
  slots.GameRecord:
    properties:
      bet:
        description: Ставка
        type: number
      result:
        description: Результат игры
        type: string
      timestamp:
        description: Время игры
        type: string
      wallet:
        description: Кошелек игрока
        type: string
      win_amount:
        description: Выигрыш
        type: number
    type: object
  slots.InitializeBalanceRequest:
    properties:
      cubes:
        description: Баланс в кубах
        type: number
      tons:
        description: Баланс в тоннах
        type: number
    type: object
  slots.PlaySlotRequest:
    properties:
      cubes:
        description: Ставка в кубах (опционально)
        type: integer
      ton:
        description: Ставка в тоннах (опционально)
        type: number
      wallet:
        description: Кошелек пользователя
        type: string
    required:
    - wallet
    type: object
  slots.PlaySlotResponse:
    properties:
      result_combo:
        description: Комбинация чисел
        items:
          type: integer
        type: array
      win_amount:
        description: Выигрыш
        type: number
    type: object
  slots.RecordGameRequest:
    properties:
      bet:
        description: Ставка игрока
        type: number
      result:
        description: Результат игры
        type: string
      wallet:
        description: Кошелек игрока
        type: string
      win_amount:
        description: Выигрыш игрока
        type: number
    type: object
  slots.SlotsBalanceResponse:
    properties:
      cubes:
        description: Баланс в кубах
        type: number
      tons:
        description: Баланс в тоннах
        type: number
      updated_at:
        description: Время последнего обновления
        type: string
    type: object
  slots.SuccessResponse:
    properties:
      status:
        description: Статус успешного выполнения
        type: string
    type: object
  slots.TokenOperationRequest:
    properties:
      amount:
        description: Сумма токенов
        type: number
      token_type:
        description: Тип токенов (tons или cubes)
        type: string
    required:
    - amount
    - token_type
    type: object
  slots.UpdateBalanceRequest:
    properties:
      cubes_delta:
        description: Изменение баланса в кубах
        type: number
      tons_delta:
        description: Изменение баланса в тоннах
        type: number
    type: object
host: api.m5dice.com
info:
  contact: {}
  description: API for managing users and games in TG-Dice application
  title: TG-Dice API
  version: "1.0"
paths:
...
  /slots/{wallet}/games:
    get:
      consumes:
      - application/json
      description: Получить все игры, сыгранные пользователем по его кошельку
      parameters:
      - description: Кошелек игрока
        in: path
        name: wallet
        required: true
        type: string
      - description: Лимит количества игр
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список игр
          schema:
            items:
              $ref: '#/definitions/slots.GameRecord'
            type: array
        "400":
          description: Ошибка с некорректным лимитом
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
        "500":
          description: Ошибка сервера при получении игр
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
      summary: Получить все игры по кошельку
      tags:
      - Slots
  /slots/{wallet}/recent-games:
    get:
      consumes:
      - application/json
      description: Получить последние игры пользователя по его кошельку
      parameters:
      - description: Кошелек игрока
        in: path
        name: wallet
        required: true
        type: string
      - description: Лимит количества игр
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список последних игр
          schema:
            items:
              $ref: '#/definitions/slots.GameRecord'
            type: array
        "400":
          description: Ошибка с некорректным лимитом
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
        "500":
          description: Ошибка сервера при получении последних игр
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
      summary: Получить последние игры по кошельку
      tags:
      - Slots
  /slots/balance:
    get:
      consumes:
      - application/json
      description: Возвращает общий баланс в тоннах и кубах
      produces:
      - application/json
      responses:
        "200":
          description: Общий баланс
          schema:
            $ref: '#/definitions/slots.SlotsBalanceResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
      summary: Получение общего баланса
      tags:
      - Slots
  /slots/balance/add:
    post:
      consumes:
      - application/json
      description: Добавляет указанное количество токенов к балансу
      parameters:
      - description: Данные для добавления токенов
        in: body
        name: tokenOperationRequest
        required: true
        schema:
          $ref: '#/definitions/slots.TokenOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Токены успешно добавлены
          schema:
            $ref: '#/definitions/slots.SuccessResponse'
        "400":
          description: Некорректные данные запроса
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
      summary: Добавление токенов
      tags:
      - Slots
  /slots/balance/initialize:
    post:
      consumes:
      - application/json
      description: Устанавливает общий баланс в тоннах и кубах
      parameters:
      - description: Данные для инициализации баланса
        in: body
        name: initializeBalanceRequest
        required: true
        schema:
          $ref: '#/definitions/slots.InitializeBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Баланс успешно инициализирован
          schema:
            $ref: '#/definitions/slots.SuccessResponse'
        "400":
          description: Некорректные данные запроса
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
      summary: Инициализация баланса
      tags:
      - Slots
  /slots/balance/subtract:
    post:
      consumes:
      - application/json
      description: Вычитает указанное количество токенов из баланса
      parameters:
      - description: Данные для вычитания токенов
        in: body
        name: tokenOperationRequest
        required: true
        schema:
          $ref: '#/definitions/slots.TokenOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Токены успешно вычтены
          schema:
            $ref: '#/definitions/slots.SuccessResponse'
        "400":
          description: Некорректные данные запроса
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
      summary: Вычитание токенов
      tags:
      - Slots
  /slots/balance/update:
    patch:
      consumes:
      - application/json
      description: Изменяет общий баланс в тоннах и кубах
      parameters:
      - description: Данные для обновления баланса
        in: body
        name: updateBalanceRequest
        required: true
        schema:
          $ref: '#/definitions/slots.UpdateBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Баланс успешно обновлен
          schema:
            $ref: '#/definitions/slots.SuccessResponse'
        "400":
          description: Некорректные данные запроса
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
      summary: Обновление баланса
      tags:
      - Slots
  /slots/play:
    post:
      consumes:
      - application/json
      description: Выполнение игры в слоты с выбором ставки
      parameters:
      - description: Параметры игры
        in: body
        name: playSlotRequest
        required: true
        schema:
          $ref: '#/definitions/slots.PlaySlotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Результат игры
          schema:
            $ref: '#/definitions/slots.PlaySlotResponse'
        "400":
          description: Ошибка с некорректной ставкой
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
      summary: Запуск игры в слоты
      tags:
      - Slots
  /slots/record:
    post:
      consumes:
      - application/json
      description: Сохранить информацию о сыгранной игре в базе данных
      parameters:
      - description: Данные игры
        in: body
        name: recordGameRequest
        required: true
        schema:
          $ref: '#/definitions/slots.RecordGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Игра успешно записана
          schema:
            $ref: '#/definitions/slots.SuccessResponse'
        "400":
          description: Ошибка с некорректной ставкой или выигрышем
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
        "500":
          description: Ошибка сервера при записи игры
          schema:
            $ref: '#/definitions/slots.ErrorResponse'
      summary: Записать информацию о сыгранной игре
      tags:
      - Slots
...
schemes:
- https
swagger: "2.0"